Added scout support for index update
Update EloquentRepository.php

+++ src/EloquentRepository.php
@@ -52,16 +52,55 @@ abstract class EloquentRepository implements EloquentRepositoryInterface
             return 0;
         }
 
-        return $this->getModel()->upsert($values, $uniqueBy, $update ?? []);
+        if (empty($uniqueBy)) {
+           return 0;
+        }
+
+        $result = $this->getModel()->upsert($values, $uniqueBy, $update ?? []);
+
+        if ($result > 0 && $this->isHasScoutSearchableTrait()){
+            $whereValues = collect($values)->map(function ($item) use  ($uniqueBy) {
+                $values = [];
+                $uniqueByValues = (is_array($uniqueBy)) ? $uniqueBy : [$uniqueBy];
+                foreach ($uniqueByValues as $value) {
+                    $itemValue = $item[$value] ?? null;
+                    if (is_null($itemValue)) {
+                        continue;
+                    }
+
+                    $values[] = $itemValue;
+                }
+
+                return $values;
+            })->toArray();
+
+            $this->query()->whereIn($uniqueBy, $whereValues)->searchable();
+        }
+
+        return $result;
     }
 
     public function update(int $id, array $data): bool
     {
+        if ($id < 1) {
+            return false;
+        }
+
         if (empty($data)) {
             return false;
         }
 
-        return $this->getModel()->where('id', $id)->update($data) > 0;
+        $result = $this->getModel()->where('id', $id)->update($data) > 0;
+
+        if (!$result) {
+            return false;
+        }
+
+        if ($this->isHasScoutSearchableTrait()) {
+            $this->find($id)?->searchable();
+        }
+
+        return true;
     }
 
     public function exists(int $id): bool
@@ -76,11 +115,27 @@ abstract class EloquentRepository implements EloquentRepositoryInterface
 
     public function delete(int $id): bool
     {
-        return $this->getModel()->whereKey($id)->delete() > 0;
+        if ($id < 1) {
+            return false;
+        }
+
+        if (!$this->isHasScoutSearchableTrait()) {
+            return $this->getModel()->whereKey($id)->delete() > 0;
+        }
+
+        return $this->find($id)?->delete() ?? false;
     }
 
     public function query(): Builder
     {
         return $this->getModel()->newQuery();
+    }
+
+    private function isHasScoutSearchableTrait(): bool
+    {
+        return in_array(
+            \Laravel\Scout\Searchable::class,
+            class_uses_recursive($this->class)
+        );
     }
 }
